+++ mex/editor/client/src/app/fast-track/create-item-dialog/create-contact-point-form/create-contact-point-form.ts
@@ -28,7 +28,7 @@ export class CreateContactPointForm {
     () => this.data() ?? this.createEmptyCreateContactPoint(this.inputText() ?? ""),
   );
 
-  contactPointSchema = CreateContactPointSchema;
+  protected readonly contactPointSchema = CreateContactPointSchema;
 
   protected readonly contactPointForm = form(this.formModel, (schema) => {
     validateStandardSchema(schema, CreateContactPointSchema);