+++ mex/editor/client/src/app/fast-track/reference-select/reference-select.ts
@@ -134,7 +134,7 @@ export class ReferenceSelect<
     }
 
     return {
-      url: "/api/v0/backend/preview-item",
+      url: "api/v0/backend/preview-item",
       params,
     };
   });