+++ CHANGELOG.md
@@ -11,6 +11,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## [4.1.1] - 2026-08-20
+
+### Added
+
+- added github workflow to sync to opencode
+
+### Fixed
+
+- removed leading / in reference-select search result fetching
+- align linting workflow with mex-template
+
+## [4.1.0] - 2026-08-18
+
+### Changes
+
 - new template https://github.com/robert-koch-institut/mex-template/releases/tag/1.5.0
 - renovate config changes for better angular experience
   - trying renovate dependency dashboard and check usage
@@ -19,16 +42,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ConceptOptions renamed to ConecptLookups and now use the new ToLookupPipe (which uses the LabelPipe, moved all lookup and label creation to spefic pipes).
 - Grouped everything into fasttrack folder to meet the feature approach. Some components might be moved to shared
 - added sorting linters for imports in typescript files and angular component import parameter
-
-### Deprecated
-
-### Removed
+- updated dependencies
 
 ### Fixed
 
 - manually fixed renovatebot lock file stuff
-
-### Security
 
 ## [4.0.2] - 2026-08-05